be able to search translated categories
